Conneted the endpoint to the places list, also created the database l…

+++ presentation/src/main/java/com/vehiclecompanion/composables/components/PlaceDetailsSheet.kt
@@ -183,7 +183,7 @@ fun PlaceDetailsSheet(
 }
 
 @Composable
-private fun RatingDisplay(rating: Int) {
+private fun RatingDisplay(rating: Double) {
     Row(
         verticalAlignment = Alignment.CenterVertically
     ) {
